+++ rust/numaflow-core/src/monovertex.rs
@@ -1,10 +1,11 @@
 use crate::config::{config, SDKConfig};
+use crate::error;
 use crate::metrics::MetricsState;
-use crate::shared::create_rpc_channel;
+use crate::shared::utils;
+use crate::shared::utils::create_rpc_channel;
 use crate::sink::user_defined::SinkWriter;
 use crate::source::user_defined::{SourceAcker, SourceReader};
 use crate::transformer::user_defined::SourceTransformer;
-use crate::{error, startup};
 use forwarder::ForwarderBuilder;
 use sink_pb::sink_client::SinkClient;
 use source_pb::source_client::SourceClient;
@@ -77,7 +78,7 @@ async fn shutdown_signal() {
 
 async fn start_forwarder(cln_token: CancellationToken, sdk_config: SDKConfig) -> error::Result<()> {
     // make sure that we have compatibility with the server
-    startup::check_compatibility(
+    utils::check_compatibility(
         &cln_token,
         sdk_config.source_server_info_path.into(),
         sdk_config.sink_server_info_path.into(),
@@ -128,7 +129,7 @@ async fn start_forwarder(cln_token: CancellationToken, sdk_config: SDKConfig) ->
     };
 
     // readiness check for all the ud containers
-    startup::wait_until_ready(
+    utils::wait_until_ready(
         cln_token.clone(),
         &mut source_grpc_client,
         &mut sink_grpc_client,
@@ -149,10 +150,10 @@ async fn start_forwarder(cln_token: CancellationToken, sdk_config: SDKConfig) ->
 
     // start the metrics server
     // FIXME: what to do with the handle
-    startup::start_metrics_server(metrics_state).await;
+    utils::start_metrics_server(metrics_state).await;
 
     // start the lag reader to publish lag metrics
-    let mut lag_reader = startup::create_lag_reader(source_grpc_client.clone()).await;
+    let mut lag_reader = utils::create_lag_reader(source_grpc_client.clone()).await;
     lag_reader.start().await;
 
     // build the forwarder
@@ -189,7 +190,7 @@ mod tests {
     use crate::config::SDKConfig;
     use crate::error;
     use crate::monovertex::start_forwarder;
-    use crate::server_info::ServerInfo;
+    use crate::shared::server_info::ServerInfo;
     use numaflow::source::{Message, Offset, SourceReadRequest};
     use numaflow::{sink, source};
     use std::fs::File;
